tests/wpt/web-platform-tests/pointerevents/pointerevent_auxclick_is_a_pointerevent.html

<!DOCTYPE HTML>
<title>auxclick is a PointerEvent</title>
<link rel="help" href="https://github.com/w3c/pointerevents/pull/317">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<input id="target" style="margin: 20px">

<script>
  'use strict';
  let auxclickTest = async_test("auxclick is a PointerEvent");
  let target = document.getElementById("target");

  target.addEventListener("auxclick", auxclickTest.step_func((e)=>{
    assert_true(e instanceof PointerEvent, "auxclick should be a PointerEvent");
  }));
  let eventWatcher = new EventWatcher(auxclickTest, target, ["auxclick"]);
  let actions = new test_driver.Actions();
  actions = actions.pointerMove(0,0, {origin:target})
    .pointerDown({button:actions.ButtonType.MIDDLE})
    .pointerUp({button:actions.ButtonType.MIDDLE});
  Promise.all([eventWatcher.wait_for("auxclick"), actions.send()]).then(()=>auxclickTest.done());
</script>
